+++ packages/lumx-core/src/js/utils/index.ts
@@ -1,1 +1,2 @@
+export * as classNames from './classNames';
 export * from './events';