+++ packages/lumx-react/src/components/message/Message.tsx
@@ -1,12 +1,11 @@
 import { ReactNode } from 'react';
 
-import classNames from 'classnames';
-
 import { mdiAlert, mdiAlertCircle, mdiCheckCircle, mdiClose, mdiInformation } from '@lumx/icons';
 import { ColorPalette, Emphasis, Icon, IconButton, Kind, Size } from '@lumx/react';
 import { GenericProps } from '@lumx/react/utils/type';
 import { handleBasicClasses } from '@lumx/core/js/utils/_internal/className';
 import type { LumxClassName } from '@lumx/core/js/types';
+import { classNames } from '@lumx/core/js/utils';
 import { forwardRef } from '@lumx/react/utils/react/forwardRef';
 
 /**
@@ -70,7 +69,7 @@ export const Message = forwardRef<MessageProps, HTMLDivElement>((props, ref) => 
     return (
         <div
             ref={ref}
-            className={classNames(
+            className={classNames.join(
                 className,
                 handleBasicClasses({
                     color,