+++ packages/lumx-react/src/components/progress/Progress.tsx
@@ -1,9 +1,8 @@
-import classNames from 'classnames';
-
 import { Theme } from '@lumx/react';
 import { GenericProps, HasTheme, ValueOf } from '@lumx/react/utils/type';
 import { handleBasicClasses } from '@lumx/core/js/utils/_internal/className';
 import type { LumxClassName } from '@lumx/core/js/types';
+import { classNames } from '@lumx/core/js/utils';
 import { useTheme } from '@lumx/react/utils/theme/ThemeContext';
 import { forwardRef } from '@lumx/react/utils/react/forwardRef';
 
@@ -57,7 +56,7 @@ export const Progress = forwardRef<ProgressProps, HTMLDivElement>((props, ref) =
         <div
             ref={ref}
             {...forwardedProps}
-            className={classNames(className, handleBasicClasses({ prefix: CLASSNAME, theme, variant }))}
+            className={classNames.join(className, handleBasicClasses({ prefix: CLASSNAME, theme, variant }))}
         >
             {variant === ProgressVariant.circular && <ProgressCircular theme={theme} />}
             {variant === ProgressVariant.linear && <ProgressLinear theme={theme} />}