+++ packages/lumx-react/src/components/radio-button/RadioGroup.tsx
@@ -1,9 +1,8 @@
 import { ReactNode } from 'react';
 
-import classNames from 'classnames';
-
 import { GenericProps } from '@lumx/react/utils/type';
 import type { LumxClassName } from '@lumx/core/js/types';
+import { classNames } from '@lumx/core/js/utils';
 import { forwardRef } from '@lumx/react/utils/react/forwardRef';
 
 /**
@@ -35,7 +34,7 @@ export const RadioGroup = forwardRef<RadioGroupProps, HTMLDivElement>((props, re
     const { children, className, ...forwardedProps } = props;
 
     return (
-        <div ref={ref} {...forwardedProps} className={classNames(className, CLASSNAME)}>
+        <div ref={ref} {...forwardedProps} className={classNames.join(className, CLASSNAME)}>
             {children}
         </div>
     );