+++ packages/lumx-react/src/components/skeleton/SkeletonCircle.tsx
@@ -1,9 +1,8 @@
-import classNames from 'classnames';
-
 import { GlobalSize, Theme, ColorPalette } from '@lumx/react';
 import { GenericProps, HasTheme } from '@lumx/react/utils/type';
 import { handleBasicClasses } from '@lumx/core/js/utils/_internal/className';
 import type { LumxClassName } from '@lumx/core/js/types';
+import { classNames } from '@lumx/core/js/utils';
 import { useTheme } from '@lumx/react/utils/theme/ThemeContext';
 import { forwardRef } from '@lumx/react/utils/react/forwardRef';
 
@@ -44,7 +43,7 @@ export const SkeletonCircle = forwardRef<SkeletonCircleProps, HTMLDivElement>((p
         <div
             ref={ref}
             {...forwardedProps}
-            className={classNames(className, handleBasicClasses({ prefix: CLASSNAME, size, color, theme }))}
+            className={classNames.join(className, handleBasicClasses({ prefix: CLASSNAME, size, color, theme }))}
         />
     );
 });